+++ app/src/main/java/cc/rits/openhacku2022/repository/TransactionRepository.java
@@ -60,7 +60,8 @@ public class TransactionRepository {
      * @return 取引リスト
      */
     public List<TransactionModel> selectAll() {
-        return this.transactionMapper.selectAll().stream() //
+        final var example = new TransactionExample();
+        return this.transactionMapper.selectByExample(example).stream() //
             .map(TransactionModel::new) //
             .collect(Collectors.toList());
     }
