+++ app/src/main/java/cc/rits/openhacku2022/service/TransactionService.java
@@ -1,5 +1,7 @@
 package cc.rits.openhacku2022.service;
 
+import org.springframework.session.FindByIndexNameSessionRepository;
+import org.springframework.session.Session;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,6 +13,7 @@ import cc.rits.openhacku2022.model.TransactionModel;
 import cc.rits.openhacku2022.query_service.TransactionQueryService;
 import cc.rits.openhacku2022.query_service.dto.TransactionWithOrderDto;
 import cc.rits.openhacku2022.repository.OrderRepository;
+import cc.rits.openhacku2022.repository.TransactionRepository;
 import lombok.RequiredArgsConstructor;
 
 /**
@@ -24,6 +27,10 @@ public class TransactionService {
     private final TransactionQueryService transactionQueryService;
 
     private final OrderRepository orderRepository;
+
+    private final TransactionRepository transactionRepository;
+
+    private final FindByIndexNameSessionRepository<? extends Session> sessionRepository;
 
     private final CheckoutClient checkoutClient;
 
@@ -65,6 +72,16 @@ public class TransactionService {
             .sum();
 
         return checkoutClient.send(paymentAmount);
+    }
+
+    /**
+     * 期限切れの取引を削除
+     */
+    public void deleteExpiredTransaction() {
+        final var transactions = this.transactionRepository.selectAll();
+        transactions.stream() //
+            .filter(transaction -> this.sessionRepository.findByPrincipalName(transaction.getCode()).isEmpty()) //
+            .forEach(this.transactionRepository::delete);
     }
 
 }
